backend/pyproject.toml
[project]
name = "vitaguard-backend"
version = "1.0.0"
description = "VitaGuard — Secure healthcare backend with AI-powered X-ray analysis"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "aiosqlite>=0.20",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.18",
    "Pillow>=11.0",
    "numpy>=2.0",
    "slowapi>=0.1",
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.9",
]
prod = [
    "asyncpg>=0.30",
]

[tool.setuptools.packages.find]
include = ["app", "app.*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "SIM"]
ignore = ["S105", "S106"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
